add charts to mobile app

+++ backend/app/models.py
@@ -200,7 +200,7 @@ class TrackingStatus(PaginatedAPIMixin, db.Model):
     condition_id = db.Column(db.Integer, db.ForeignKey('condition.id'))
     product_id = db.Column(db.Integer, db.ForeignKey('product.id'))
     tracking_device_id = db.Column(db.Integer, db.ForeignKey('tracking_device.id'))
-    value = db.Column(db.Integer)
+    value = db.Column(db.Float)
     date_recordered = db.Column(db.DateTime)
 
     @staticmethod
